Begin implementing field post-processing
